[refac] template stage/container moves to container/task
Enables:
- stage to container
- stage to task
- container to container
- container to task
