--- include/behaviortree_mtc/move_mtc_container_to_parent_container.h
@@ -1,18 +1,0 @@
-#pragma once
-
-#include <behaviortree_cpp/bt_factory.h>
-
-namespace BT {
-namespace MTC {
-
-class MoveMTCContainerToParentContainer : public BT::SyncActionNode
-{
-public:
-  MoveMTCContainerToParentContainer(const std::string& name, const BT::NodeConfig& config);
-
-  BT::NodeStatus tick() override;
-  static BT::PortsList providedPorts();
-};
-
-}  // namespace MTC
-}  // namespace BT
