--- include/behaviortree_mtc/move_mtc_stage_to_container.h
@@ -1,18 +1,0 @@
-#pragma once
-
-#include <behaviortree_cpp/bt_factory.h>
-
-namespace BT {
-namespace MTC {
-
-class MoveMTCStageToContainer : public BT::SyncActionNode
-{
-public:
-  MoveMTCStageToContainer(const std::string& name, const BT::NodeConfig& config);
-
-  BT::NodeStatus tick() override;
-  static BT::PortsList providedPorts();
-};
-
-}  // namespace MTC
-}  // namespace BT
